# docs/swagger/swagger.yaml
basePath: /
info:
  contact: {}
  description: 在线客服API文档的文档，接管了微信公众号聊天
  license: {}
  title: 在线客服API文档
  version: 0.0.1
paths:
  /v1/customer/{customerId}/history:
    get:
      consumes:
      - application/json
      description: 获取一个用户的聊天记录
      parameters:
      - description: 客户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取一个用户的聊天记录
      tags:
      - Dialog
  /v1/customer/message:
    post:
      consumes:
      - application/json
      description: 客服发送消息给客户
      parameters:
      - description: 客户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 客服发送消息给客户
      tags:
      - Dialog
  /v1/dialog/access:
    post:
      consumes:
      - application/json
      description: 客服接入用户，客服加入会话房间
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 客服接入用户，客服加入会话房间
      tags:
      - Dialog
  /v1/dialog/ack:
    put:
      consumes:
      - application/json
      description: 确认已读用户的消息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 确认已读用户的消息
      tags:
      - Dialog
  /v1/dialog/list:
    get:
      consumes:
      - application/json
      description: 获取待回复消息列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取待回复消息列表
      tags:
      - Dialog
  /v1/health:
    get:
      consumes:
      - application/json
      description: 应用程序健康检查接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 健康检查
      tags:
      - Default
  /v1/init:
    get:
      consumes:
      - application/json
      description: 在线客服系统进行初始化
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 系统初始化
      tags:
      - Default
  /v1/kf/{kfId}:
    get:
      consumes:
      - application/json
      description: 获取客服信息
      parameters:
      - description: 客服的ID
        in: path
        name: kfId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取客服信息
      tags:
      - Kf
  /v1/kf/{kfId}/status:
    put:
      consumes:
      - application/json
      description: 客服修改在线状态
      parameters:
      - description: 客服的ID
        in: path
        name: kfId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 客服修改在线状态
      tags:
      - Kf
  /v1/login/{tokenId}:
    post:
      consumes:
      - application/json
      description: 客服登入
      parameters:
      - description: 客服的授权TokenId
        in: path
        name: tokenId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"msg":"ok"}'
          schema:
            type: string
      summary: 客服登入
      tags:
      - Kf
  /v1/setting/offline_reply:
    get:
      consumes:
      - application/json
      description: 获取自动回复列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取自动回复列表
      tags:
      - OfflineReply
    post:
      consumes:
      - application/json
      description: 新增一条离线自动回复语句
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 新增一条离线自动回复语句
      tags:
      - OfflineReply
  /v1/setting/offline_reply/{id}:
    delete:
      consumes:
      - application/json
      description: 删除一条离线自动回复语句
      parameters:
      - description: 自动回复语句的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 删除一条离线自动回复语句
      tags:
      - OfflineReply
    put:
      consumes:
      - application/json
      description: 更新一条离线自动回复语句
      parameters:
      - description: 自动回复语句的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 更新一条离线自动回复语句
      tags:
      - OfflineReply
  /v1/wait_queue/:
    get:
      consumes:
      - application/json
      description: 待接入列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 待接入列表
      tags:
      - WaitQueue
swagger: "2.0"